refactor(web): browser-based subkey array init + display management
