+++ web/source/osk/visualKeyboard.ts
@@ -2,6 +2,7 @@
 /// <reference path="utils.ts" />
 /// <reference path="abstractions.ts" />
 /// <reference path="browser/keytip.ts" />
+/// <reference path="browser/subkeyPopup.ts" />
 
 namespace com.keyman.osk {
   let Codes = com.keyman.text.Codes;
@@ -673,7 +674,7 @@ namespace com.keyman.osk {
     popupDelay: number = 500;
     menuEvent: KeyElement; // Used by embedded-mode.
     keytip: KeyTip;
-    popupCallout: HTMLDivElement;
+    subkeyPopup: browser.SubkeyPopup;
 
     get layerId(): string {
       return this._layerId;
@@ -1660,20 +1661,11 @@ namespace com.keyman.osk {
     }
 
     clearPopup() {
-      let oskManager = com.keyman.singleton.osk;
       // Remove the displayed subkey array, if any, and cancel popup request
-      var sk=document.getElementById('kmw-popup-keys');
-      if(sk != null) {
-        if(sk.shim) {
-          oskManager._Box.removeChild(sk.shim);
-        }
-        sk.parentNode.removeChild(sk);
+      if(this.subkeyPopup) {
+        this.subkeyPopup.clear();
+        this.subkeyPopup = null;
       }
-
-      if(this.popupCallout) {
-        oskManager._Box.removeChild(this.popupCallout);
-      }
-      this.popupCallout = null;
 
       if(this.subkeyDelayTimer) {
           window.clearTimeout(this.subkeyDelayTimer);
@@ -1693,143 +1685,20 @@ namespace com.keyman.osk {
         return;
       }
 
-      let keyman = com.keyman.singleton;
-      let device = this.device;
-
-      // A tag we directly set on a key element during its construction.
-      let subKeySpec: OSKKeySpec[] = e['subKeys'];
-
-      // Create holder DIV for subkey array, and set styles.
-      // A subkey array for Shift will only appear if extra layers exist
-
-      // The holder is position:fixed, but the keys do not need to be, as no scrolling
-      // is possible while the array is visible.  So it is simplest to let the keys have
-      // position:static and display:inline-block
-      var subKeys=document.createElement('DIV'),i;
-
-      subKeys.id='kmw-popup-keys';
-      this.popupBaseKey = e;
-
-      // #3718: No longer prepend base key to popup array
-
-      // Must set position dynamically, not in CSS
-      var ss=subKeys.style;
-      ss.bottom=(parseInt(e.style.bottom,10)+parseInt(e.style.height,10)+4)+'px';
-
-      // Set key font according to layout, or defaulting to OSK font
-      // (copied, not inherited, since OSK is not a parent of popup keys)
-      ss.fontFamily=this.fontFamily;
-
-      // Copy the font size from the parent key, allowing for style inheritance
-      ss.fontSize=keyman.util.getStyleValue(e,'font-size');
-      ss.visibility='hidden';
-
-      var nKeys=subKeySpec.length,nRow,nRows,nCols;
-      nRows=Math.min(Math.ceil(nKeys/9),2);
-      nCols=Math.ceil(nKeys/nRows);
-      if(nRows > 1) {
-        ss.width=(nCols*e.offsetWidth+nCols*5)+'px';
-      }
-
-      // Add nested button elements for each sub-key
-      for(i=0; i<nKeys; i++) {
-        var needsTopMargin = false;
-        let nRow=Math.floor(i/nCols);
-        if(nRows > 1 && nRow > 0) {
-          needsTopMargin = true;
-        }
-
-        let layer = e['key'].layer;
-        if(typeof(layer) != 'string' || layer == '') {
-          // Use the currently-active layer.
-          layer = this.layerId;
-        }
-        let keyGenerator = new com.keyman.osk.OSKSubKey(subKeySpec[i], layer);
-        let kDiv = keyGenerator.construct(this, <KeyElement> e, needsTopMargin);
-
-        subKeys.appendChild(kDiv);
-      }
-
       // Clear key preview if any
       this.showKeyTip(null,false);
+      this.popupBaseKey = e;
+
+      let subKeys = this.subkeyPopup = new browser.SubkeyPopup(this, e);
 
       // Otherwise append the touch-hold (subkey) array to the OSK
-      keyman.osk._Box.appendChild(subKeys);
-
-      // And correct its position with respect to that element
-      ss=subKeys.style;
-      var x = dom.Utils.getAbsoluteX(e)+0.5*(e.offsetWidth-subKeys.offsetWidth);
-      var xMax = keyman.osk.getWidth() - subKeys.offsetWidth;
-
-      if(x > xMax) {
-        x=xMax;
-      }
-      if(x < 0) {
-        x=0;
-      }
-      ss.left=x+'px';
-
-      // Make the popup keys visible
-      ss.visibility='visible';
-
-      // For now, should only be true (in production) when keyman.isEmbedded == true.
-      let constrainPopup = keyman.isEmbedded;
-
-      let cs = getComputedStyle(subKeys);
-      let oskHeight = keyman.osk.getHeight();
-      let bottomY = parseInt(cs.bottom, 10);
-      let popupHeight = parseInt(cs.height, 10);
-
-      let delta = 0;
-      if(popupHeight + bottomY > oskHeight && constrainPopup) {
-        delta = popupHeight + bottomY - oskHeight;
-        ss.bottom = (bottomY - delta) + 'px';
-      }
-
-      // Add the callout
-      this.popupCallout = this.addCallout(e, delta);
-
-      // And add a filter to fade main keyboard
-      subKeys.shim = document.createElement('DIV');
-      subKeys.shim.id = 'kmw-popup-shim';
+      let keyman = com.keyman.singleton;
+      keyman.osk._Box.appendChild(subKeys.element);
       keyman.osk._Box.appendChild(subKeys.shim);
 
-      // Highlight the duplicated base key or ideal subkey (if a phone)
-      if(device.formFactor == 'phone') {
-        this.selectDefaultSubkey(e, subKeySpec, subKeys);
-      }
+      // Must be placed after its `.element` has been inserted into the DOM.
+      subKeys.reposition(this);
     }
-
-    selectDefaultSubkey(baseKey: KeyElement, subkeys: OSKKeySpec[], popupBase: HTMLElement) {
-      var bk: KeyElement;
-      for(let i=0; i < subkeys.length; i++) {
-        let skSpec = subkeys[i];
-        let skElement = <KeyElement> popupBase.childNodes[i].firstChild;
-
-        // Preference order:
-        // #1:  if a default subkey has been specified, select it.  (pending, for 15.0+)
-        // #2:  if no default subkey is specified, default to a subkey with the same 
-        //      key ID and layer / modifier spec.
-        //if(skSpec.isDefault) { TODO for 15.0
-        //  bk = skElement;
-        //  break;
-        //} else
-        if(!baseKey.key || !baseKey.key.spec) {
-          continue;
-        }
-
-        if(skSpec.elementID == baseKey.key.spec.elementID) {
-          bk = skElement;
-          break; // Best possible match has been found.  (Disable 'break' once above block is implemented.)
-        }
-      }
-
-      if(bk) {
-        this.keyPending = bk;
-        this.highlightKey(bk,true);//bk.className = bk.className+' kmw-key-touched';
-      }
-    }
-
 
     //#endregion
 
@@ -2537,45 +2406,6 @@ namespace com.keyman.osk {
         }
       }
     };
-
-    /**
-     * Add a callout for popup keys (if KeymanWeb on a phone device)
-     *
-     * @param   {Object}  key   HTML key element
-     * @return  {Object}        callout object
-     */
-    addCallout(key: KeyElement, delta?: number): HTMLDivElement {
-      let keyman = com.keyman.singleton;
-
-      if(this.device.formFactor != 'phone' || this.device.OS != 'iOS') {
-        return null;
-      }
-
-      delta = delta || 0;
-
-      let calloutHeight = key.offsetHeight - delta;
-
-      if(calloutHeight > 0) {
-        var cc = document.createElement('div'), ccs = cc.style;
-        cc.id = 'kmw-popup-callout';
-        keyman.osk._Box.appendChild(cc);
-
-        // Create the callout
-        var xLeft = key.offsetLeft,
-            xTop = key.offsetTop + delta,
-            xWidth = key.offsetWidth,
-            xHeight = calloutHeight;
-
-        // Set position and style
-        ccs.top = (xTop-6)+'px'; ccs.left = xLeft+'px';
-        ccs.width = xWidth+'px'; ccs.height = (xHeight+6)+'px';
-
-        // Return callout element, to allow removal later
-        return cc;
-      } else {
-        return null;
-      }
-    }
 
     /**
      * Wait until font is loaded before applying stylesheet - test each 100 ms
